backend/win64-py39-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=win64-py39-requirements.txt
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
cachetools==6.2.0
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
exceptiongroup==1.3.0
    # via anyio
fastapi[standard]==0.118.0
    # via -r requirements.in
fastapi-cli[standard]==0.0.13
    # via fastapi
fastapi-cloud-cli==0.3.0
    # via fastapi-cli
google-auth==2.41.1
    # via google-genai
google-genai==1.41.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic[email]==2.11.10
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pymongo==4.15.2
    # via -r requirements.in
python-dotenv==1.1.1
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via google-genai
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.0
    # via fastapi-cloud-cli
rsa==4.9.1
    # via google-auth
sentry-sdk==2.39.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via fastapi
tenacity==9.1.2
    # via google-genai
typer==0.19.2
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.37.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
